# Conflicts:
#	gradle.properties
#	surf-shop-paper/src/main/kotlin/dev/slne/surf/shop/paper/menu/ShopListView.kt
#	surf-shop-paper/src/main/kotlin/dev/slne/surf/shop/paper/menu/edit/storage/ItemStorageInsertView.kt
#	surf-shop-paper/src/main/kotlin/dev/slne/surf/shop/paper/menu/menu-utils.kt

+++ surf-shop-microservice/src/main/kotlin/dev/slne/surf/shop/backend/repository/ShopRepository.kt
@@ -27,7 +27,7 @@ class ShopRepository {
     suspend fun createShop(
         itemString: String,
         storedItemCount: Int,
-        pricePerItem: Int,
+        pricePerItem: Double,
         seller: UUID,
         createdAt: OffsetDateTime
     ): Shop = suspendTransaction {
